feat: automatic git auth and signing for sprout agents
Integrate git-credential-nostr and git-sign-nostr into sprout-dev-mcp as
multicall personalities. Agents using sprout-dev-mcp get automatic NIP-98
git auth and NIP-GS commit signing when NOSTR_PRIVATE_KEY is set — zero
persistent config mutation.

## How it works

sprout-dev-mcp is already a multicall binary (rg, tree, sprout CLI via
argv[0] dispatch). This adds git-credential-nostr and git-sign-nostr as
two more personalities, with a shim directory that puts them on PATH.

At startup, the dev-mcp:
1. Reads NOSTR_PRIVATE_KEY, then removes it from the process env
2. Writes the key to a 0600 file (created with O_CREAT|O_EXCL + mode)
3. Derives the public key for user.signingkey
4. Zeroizes the in-memory key string
5. Builds ephemeral GIT_CONFIG_* env vars for shell children

Shell children get automatic nostr credential auth + commit signing via
the ephemeral git config. The credential helper is additive — it silently
declines non-Sprout remotes (no Nostr WWW-Authenticate challenge = exit 0,
no credential), so git falls through to system helpers for GitHub/GitLab.

## Changes

- git-credential-nostr: extract lib.rs with `pub fn run() -> i32`
- git-sign-nostr: extract lib.rs with `pub fn run() -> i32`
- sprout-dev-mcp: multicall dispatch (sync, before tokio/tracing),
  shim symlinks, keyfile, ephemeral git config, secret scoping
- sprout-agent: widen PASSTHROUGH_ENV (SSH_AUTH_SOCK, GIT_ASKPASS,
  GIT_SSH_COMMAND, SPROUT_PRIVATE_KEY, SPROUT_RELAY_URL, etc.)
- scripts/build-agent-release.sh: tar packaging with cross-compile support
- Remove unused anyhow dep from git-credential-nostr

Standalone binaries for both git helpers continue to work unchanged.
