+++ crates/sprout-dev-mcp/Cargo.toml
@@ -11,6 +11,10 @@ path = "src/main.rs"
 
 [dependencies]
 sprout-cli = { path = "../sprout-cli" }
+git-credential-nostr = { path = "../git-credential-nostr" }
+git-sign-nostr = { path = "../git-sign-nostr" }
+nostr = { workspace = true }
+zeroize = { workspace = true }
 tokio = { workspace = true }
 tokio-util = { workspace = true }
 serde = { workspace = true }
